Start to reflect UX mockup - Move non-tab components out of tabbar, move navbar above tabbar, make navbar drag the entire electron app, reorder the icons in navbar, move the OS system commands to Windows/right hand style, make the URL bar its own component
