+++ static/browser.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="./css/browser.css">
     <link rel="stylesheet" href="./css/browser-tabs.css">
     <link rel="stylesheet" href="./css/browser-navbar.css">
+    <link rel="stylesheet" href="./css/browser-navbar-location.css">
     <link rel="stylesheet" href="./css/browser-page.css">
   </head>
   <body>
