+++ static/css/browser-navbar-location.css
@@ -1,0 +1,10 @@
+#browser-location-bar {
+  flex: 1;
+  display: flex;
+  margin: 0 5px;
+}
+
+#browser-location-bar input {
+  flex: 1;
+  display: flex;
+}
