+++ static/css/browser-navbar.css
@@ -8,9 +8,17 @@
   font-weight: 100;
   background: linear-gradient(to bottom, #eee, #ddd);
   border-bottom: 1px solid #aaa;
+
+  /* Makes electron app draggable */
+  -webkit-app-region: drag;
 }
 
 #browser-navbar a {
+  /* We should possibly make buttons in the navbar
+   * NOT drag the entire electron app window, but just leaving
+   * this here for now
+  -webkit-app-region: no-drag;
+  */
   text-decoration: none;
   color: #777;
   cursor: pointer;
@@ -41,29 +49,6 @@
   flex: 1;
   display: flex;
   margin: 0 5px;
-}
-
-#browser-navbar .input-group input {
-  flex: 1;
-  margin: 4px 0 3px;
-  padding: 0 5px;
-  font-size: 14px;
-  color: #808080;
-  outline: 0;
-  box-shadow: inset 0px 1px 2px rgba(0, 0, 0, 0.2);
-  background: #fff;
-  border: 1px solid #bbb;
-  border-top-left-radius: 3px;
-  border-bottom-left-radius: 3px;
-}
-
-#browser-navbar .input-group input:focus, 
-#browser-navbar .input-group input:active, 
-#browser-navbar .input-group input:hover {
-  outline: 0;
-  box-shadow: inset 0px 1px 2px rgba(0, 0, 0, 0.2);
-  background: #fff;
-  border: 1px solid #bbb;
 }
 
 #browser-navbar .input-group a {
